Index map is now full screen. Add ability to view and create incidents directly on map.
On the index page, the map is now full screen and displays all incidents. The incidents
are no longer listed. Users are now able to view incident information directly on the
map by clicking on the incident. Users are also now able to create new incidents directly
on the map.